fix: add missing die
If there is just one die and roll modifier is +1 or higher, then always should add another die to the pool to create two dice pool. So, whenever `mod` is > 0 and we have one die we should just add another D6 die and decrease `mod` by 1. If there are two dice and `mod` is positive, then proceed normally.